feat(authorization): update authorize device screen part 1
