--- packages/blockchain-wallet-v4-frontend/src/scenes/AuthorizeLogin/selectors.ts
@@ -1,3 +1,0 @@
-import { selectors } from 'data'
-
-export const getData = selectors.core.data.misc.authorizeLogin
